refactor SortingAlgorithm
