+++ src/main/java/pl/paweln/codility/counting/MaxCounters.java
@@ -1,32 +1,40 @@
 package pl.paweln.codility.counting;
 
-import pl.paweln.codility.core.CodilitySolution;
+import pl.paweln.codility.core.AbstractCodility;
 import pl.paweln.codility.core.SolutionInputParams;
 
 import java.util.Arrays;
 
-public class MaxCounters implements CodilitySolution {
+public class MaxCounters extends AbstractCodility {
+
+    public MaxCounters() {
+        int[] tabA = new int[] {3, 4, 4, 6, 1, 4, 4};
+
+        this.defaultParams = this.paramsBuilder
+                .setFirstArray(tabA)
+                .setN(5).build();
+    }
 
     @Override
     public int[] solution(SolutionInputParams params) {
-        int N = params.getN();
+        int n = params.getN();
         int[] tab = params.getFirstArray();
 
         if (tab.length == 0 || tab.length > 100000)
             throw new IllegalArgumentException("Array size must be within range 1 .. 100000.");
 
-        if (N < 0 || N > 100000)
+        if (n < 0 || n > 100000)
             throw new IllegalArgumentException("N must be within range 1 .. 100000.");
 
-        return this.solution(N, tab);
+        return this.solution(n, tab);
     }
 
     private int[] solution(int N, int[] A) {
         int lastMax = 0;
         int maxValue = 0;
 
-        int[] R = new int[N];
-        Arrays.fill(R, 0);
+        int[] tabR = new int[N];
+        Arrays.fill(tabR, 0);
 
         if (N <= 0) {
             throw new IllegalArgumentException("N must be higher of equal to 1.");
@@ -38,14 +46,14 @@ public class MaxCounters implements CodilitySolution {
             }
 
             if (aA < N + 1) {
-                if (R[aA - 1] < lastMax) {
-                    R[aA - 1] = lastMax;
+                if (tabR[aA - 1] < lastMax) {
+                    tabR[aA - 1] = lastMax;
                 }
 
-                R[aA - 1]++;
+                tabR[aA - 1]++;
 
-                if (R[aA - 1] > maxValue) {
-                    maxValue = R[aA - 1];
+                if (tabR[aA - 1] > maxValue) {
+                    maxValue = tabR[aA - 1];
                 }
 
             } else {
@@ -53,14 +61,14 @@ public class MaxCounters implements CodilitySolution {
             }
         }
 
-        for (int i = 0; i < R.length; i++) {
-            if (R[i] < lastMax) {
-                R[i] = lastMax;
+        for (int i = 0; i < tabR.length; i++) {
+            if (tabR[i] < lastMax) {
+                tabR[i] = lastMax;
             }
 
         }
 
-        return R;
+        return tabR;
     }
