+++ src/test/java/pl/paweln/codility/counting/MaxCountersTest.java
@@ -10,41 +10,41 @@ import java.util.Arrays;
 public class MaxCountersTest {
     private final SolutionFactory solutionFactory = new SolutionFactory();
 
-    private CodilitySolution s;
+    private CodilitySolution codilitySolution;
     private SolutionInputParams.SolutionInputParamsBuilder paramsBuilder;
     private SolutionInputParams params;
 
     @Before
-    public void setUp() {
-        this.s = this.solutionFactory.getSolution(SolutionFactory.CodilityTask.COUNTING_MAX_COUNTERS);
+    public void setup() {
+        this.codilitySolution = this.solutionFactory.getSolution(SolutionFactory.CodilityTask.COUNTING_MAX_COUNTERS);
         this.paramsBuilder = new SolutionInputParams.SolutionInputParamsBuilder();
     }
 
     @Test
-    public void testSolutionExample() {
-        int[] A = new int[] {3, 4, 4, 6, 1, 4, 4};
+    public void testExample() {
+        int[] tabA = new int[] {3, 4, 4, 6, 1, 4, 4};
 
         this.params = this.paramsBuilder
-                .setFirstArray(A)
+                .setFirstArray(tabA)
                 .setN(5).build();
 
-        int[] retArr = this.s.solution(this.params);
+        int[] retArr = this.codilitySolution.solution(this.params);
 
         Assert.assertArrayEquals(new int[] {3, 2, 2, 4, 2}, retArr);
     }
 
     @Test
     public void testMaxValues() {
-        int[] A = new int[100000];
+        int[] tabA = new int[100000];
 
-        Arrays.fill(A, 100000);
-        A[A.length-1] = 100001;
+        Arrays.fill(tabA, 100000);
+        tabA[tabA.length-1] = 100001;
 
         this.params = this.paramsBuilder
-                .setFirstArray(A)
+                .setFirstArray(tabA)
                 .setN(100000).build();
 
-        int[] retArr = this.s.solution(this.params);
+        int[] retArr = this.codilitySolution.solution(this.params);
 
         int[] expArr = new int[100000];
         Arrays.fill(expArr, 99999);
